requirements.txt
alembic==1.4.3
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
pytest==5.4.3
pytest-flask==1.1.0
python-dotenv==0.10.1
gunicorn==19.5.0
SQLAlchemy==1.3.17
psycopg2-binary==2.8.5
flask_sqlalchemy==2.4.4
flask_script==2.0.6
flask_migrate==2.6.0
flask_swagger==0.2.14
tox==3.21.4
flask_restplus==0.13.0
Werkzeug==0.16.1

